refactor(site/src/pages/AIBridgePage): drop getProviderIconName hack
The getProviderIconName wrapper existed to remap "anthropic" to
"anthropic-neue" so RequestLogsPage rendered the correct icon. With
that page removed, the wrapper, the FIXME comment, and the duplicate
"anthropic-neue" icon case are no longer needed. Call sites now pass
the provider directly.
