+++ site/src/pages/AIBridgePage/ListSessionsPage/ListSessionsRow.tsx
@@ -10,11 +10,11 @@ import {
 	TooltipProvider,
 	TooltipTrigger,
 } from "#/components/Tooltip/Tooltip";
-import { AIBridgeClientIcon } from "#/pages/AIBridgePage/RequestLogsPage/icons/AIBridgeClientIcon";
-import { AIBridgeProviderIcon } from "#/pages/AIBridgePage/RequestLogsPage/icons/AIBridgeProviderIcon";
+import { AIBridgeClientIcon } from "#/pages/AIBridgePage/icons/AIBridgeClientIcon";
+import { AIBridgeProviderIcon } from "#/pages/AIBridgePage/icons/AIBridgeProviderIcon";
 import { DATE_FORMAT, formatDateTime } from "#/utils/time";
 import { TokenBadges } from "../TokenBadges";
-import { getProviderDisplayName, getProviderIconName } from "../utils";
+import { getProviderDisplayName } from "../utils";
 
 type ListSessionsRowProps = {
 	session: AIBridgeSession;
@@ -71,7 +71,7 @@ export const ListSessionsRow: FC<ListSessionsRowProps> = ({
 						<Badge className="gap-1.5 max-w-full">
 							<div className="flex-shrink-0 flex items-center">
 								<AIBridgeProviderIcon
-									provider={getProviderIconName(session.providers[0])}
+									provider={session.providers[0]}
 									className="size-icon-xs"
 								/>
 							</div>
